# server/docs/swagger.yaml
basePath: /
definitions:
  api.ErrorResponse:
    properties:
      error:
        type: string
    type: object
  board.Board:
    properties:
      available_hours_per_sprint:
        type: integer
      created_at:
        type: string
      id:
        type: string
      item_seq:
        type: integer
      name:
        type: string
      sprint_cadence_days:
        type: integer
      sprint_start_day:
        type: integer
      updated_at:
        type: string
      user_id:
        type: string
    type: object
  board.CreateStatusRequest:
    properties:
      is_done:
        type: boolean
      is_initial:
        type: boolean
      name:
        type: string
      position:
        type: integer
    type: object
  board.ReorderStatusesRequest:
    properties:
      ids:
        items:
          type: string
        type: array
    type: object
  board.Status:
    properties:
      board_id:
        type: string
      created_at:
        type: string
      id:
        type: string
      is_done:
        type: boolean
      is_initial:
        type: boolean
      name:
        type: string
      position:
        type: integer
    type: object
  board.StatusListResponse:
    properties:
      statuses:
        items:
          $ref: '#/definitions/board.Status'
        type: array
    type: object
  board.UpdateRequest:
    properties:
      available_hours_per_sprint:
        type: integer
      name:
        type: string
      sprint_cadence_days:
        type: integer
      sprint_start_day:
        type: integer
    type: object
  epic.CreateRequest:
    properties:
      color:
        type: string
      deadline:
        type: string
      description:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  epic.Epic:
    properties:
      board_id:
        type: string
      color:
        type: string
      created_at:
        type: string
      deadline:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
  epic.ListResponse:
    properties:
      epics:
        items:
          $ref: '#/definitions/epic.Epic'
        type: array
    type: object
...
  epic.UpdateRequest:
    properties:
      color:
        type: string
      deadline:
        type: string
      description:
        type: string
      name:
        type: string
      type:
        type: string
    type: object
...
host: api.cumin.ashinsabu.com
info:
  contact: {}
  description: OSS sprint board API
  title: Cumin API
  version: "1.0"
paths:
  /api/board:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/board.Board'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Get board
      tags:
      - board
    patch:
      consumes:
      - application/json
      parameters:
      - description: Board settings to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/board.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/board.Board'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update board
      tags:
      - board
  /api/board/statuses:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/board.StatusListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List board statuses
      tags:
      - board
    post:
      consumes:
      - application/json
      parameters:
      - description: Status to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/board.CreateStatusRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/board.Status'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create board status
      tags:
      - board
  /api/board/statuses/{id}:
    delete:
      parameters:
      - description: Status ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "409":
          description: Conflict
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete board status
      tags:
      - board
  /api/board/statuses/reorder:
    put:
      consumes:
      - application/json
      parameters:
      - description: Ordered status IDs
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/board.ReorderStatusesRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/board.StatusListResponse'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Reorder board statuses
      tags:
      - board
...
  /api/epics:
    get:
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/epic.ListResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: List epics
      tags:
      - epics
    post:
      consumes:
      - application/json
      parameters:
      - description: Epic to create
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/epic.CreateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/epic.Epic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Create epic
      tags:
      - epics
  /api/epics/{id}:
    delete:
      parameters:
      - description: Epic ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Delete epic
      tags:
      - epics
    patch:
      consumes:
      - application/json
      parameters:
      - description: Epic ID
        in: path
        name: id
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/epic.UpdateRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/epic.Epic'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/api.ErrorResponse'
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Update epic
      tags:
      - epics
  /api/epics/{id}/restore:
    post:
      parameters:
      - description: Epic ID
        in: path
        name: id
        required: true
        type: string
      responses:
        "204":
          description: No Content
        "404":
          description: Not Found
          schema:
            $ref: '#/definitions/api.ErrorResponse'
      security:
      - CookieAuth: []
      summary: Restore epic
      tags:
      - epics
...
securityDefinitions:
  CookieAuth:
    in: cookie
    name: token
    type: apiKey
swagger: "2.0"