# server/epic/handler.go
package epic

import (
	"context"
	"time"

	"github.com/ashinsabu/cumin/server/api"
	"github.com/ashinsabu/cumin/server/auth"
	"github.com/ashinsabu/cumin/server/board"
	"github.com/go-chi/chi/v5"
)

type CreateRequest struct {
	Name        string     `json:"name"`
	Type        string     `json:"type"`
	Color       string     `json:"color"`
	Deadline    *time.Time `json:"deadline"`
	Description string     `json:"description"`
}

type UpdateRequest struct {
	Name        string     `json:"name"`
	Type        string     `json:"type"`
	Color       string     `json:"color"`
	Deadline    *time.Time `json:"deadline"`
	Description string     `json:"description"`
}

type ListResponse struct {
	Epics []Epic `json:"epics"`
}

var validTypes = map[string]bool{"recurring": true, "goal": true, "catchall": true}

type Handler struct {
	store      *Store
	boardStore *board.Store
}

func NewHandler(store *Store, boardStore *board.Store) *Handler {
	return &Handler{store: store, boardStore: boardStore}
}

func (h *Handler) Routes(r chi.Router) {
	r.Get("/api/epics", api.HandleNoBody(h.List))
	r.Post("/api/epics", api.Handle(h.Create))
	r.Patch("/api/epics/{id}", api.Handle(h.Update))
	r.Delete("/api/epics/{id}", api.HandleDelete(h.Delete))
	r.Post("/api/epics/{id}/restore", api.HandleDelete(h.Restore))
}

// List returns all epics for the authenticated user's board.
//
// @Summary      List epics
// @Tags         epics
// @Produce      json
// @Success      200  {object}  epic.ListResponse
// @Failure      404  {object}  api.ErrorResponse
// @Security     CookieAuth
// @Router       /api/epics [get]
func (h *Handler) List(ctx context.Context) (*ListResponse, error) {
	b, err := h.boardStore.GetByUser(ctx, auth.UserIDFromContext(ctx))
	if err != nil {
		return nil, api.NotFound("board not found")
	}

	epics, err := h.store.List(ctx, b.ID)
	if err != nil {
		return nil, api.Internal("failed to list epics")
	}
	if epics == nil {
		epics = []Epic{}
	}
	return &ListResponse{Epics: epics}, nil
}

// Create creates a new epic on the board.
//
// @Summary      Create epic
// @Tags         epics
// @Accept       json
// @Produce      json
// @Param        body  body  epic.CreateRequest  true  "Epic to create"
// @Success      200  {object}  epic.Epic
// @Failure      400  {object}  api.ErrorResponse
// @Failure      404  {object}  api.ErrorResponse
// @Security     CookieAuth
// @Router       /api/epics [post]
func (h *Handler) Create(ctx context.Context, req CreateRequest) (*Epic, error) {
	b, err := h.boardStore.GetByUser(ctx, auth.UserIDFromContext(ctx))
	if err != nil {
		return nil, api.NotFound("board not found")
	}

	if req.Name == "" {
		return nil, api.BadRequest("name required")
	}
	if req.Type == "" {
		req.Type = "catchall"
	}
	if !validTypes[req.Type] {
		return nil, api.BadRequest("type must be: recurring, goal, or catchall")
	}
	if req.Color == "" {
		req.Color = "#6b7280"
	}

	return h.store.Create(ctx, b.ID, req.Name, req.Type, req.Color, req.Description, req.Deadline)
}

// Update updates an existing epic.
//
// @Summary      Update epic
// @Tags         epics
// @Accept       json
// @Produce      json
// @Param        id    path  string            true  "Epic ID"
// @Param        body  body  epic.UpdateRequest  true  "Fields to update"
// @Success      200  {object}  epic.Epic
// @Failure      400  {object}  api.ErrorResponse
// @Failure      404  {object}  api.ErrorResponse
// @Security     CookieAuth
// @Router       /api/epics/{id} [patch]
func (h *Handler) Update(ctx context.Context, req UpdateRequest) (*Epic, error) {
	id := api.URLParam(ctx, "id")

	existing, err := h.store.GetByID(ctx, id)
	if err != nil {
		return nil, api.NotFound("epic not found")
	}

	b, err := h.boardStore.GetByUser(ctx, auth.UserIDFromContext(ctx))
	if err != nil || existing.BoardID != b.ID {
		return nil, api.NotFound("epic not found")
	}

	if req.Name == "" {
		req.Name = existing.Name
	}
	if req.Type == "" {
		req.Type = existing.Type
	}
	if req.Type != "" && !validTypes[req.Type] {
		return nil, api.BadRequest("type must be: recurring, goal, or catchall")
	}
	if req.Color == "" {
		req.Color = existing.Color
	}
	if req.Description == "" {
		req.Description = existing.Description
	}
	if req.Deadline == nil {
		req.Deadline = existing.Deadline
	}

	return h.store.Update(ctx, id, req.Name, req.Type, req.Color, req.Description, req.Deadline)
}

// Delete soft-deletes an epic and its child items.
//
// @Summary      Delete epic
// @Tags         epics
// @Param        id  path  string  true  "Epic ID"
// @Success      204  "No Content"
// @Failure      404  {object}  api.ErrorResponse
// @Security     CookieAuth
// @Router       /api/epics/{id} [delete]
func (h *Handler) Delete(ctx context.Context) error {
	id := api.URLParam(ctx, "id")

	existing, err := h.store.GetByID(ctx, id)
	if err != nil {
		return api.NotFound("epic not found")
	}

	b, err := h.boardStore.GetByUser(ctx, auth.UserIDFromContext(ctx))
	if err != nil || existing.BoardID != b.ID {
		return api.NotFound("epic not found")
	}

	if err := h.store.SoftDelete(ctx, id, b.ID); err != nil {
		return api.Internal("failed to delete epic")
	}
	return nil
}

// Restore restores a soft-deleted epic and its child items.
//
// @Summary      Restore epic
// @Tags         epics
// @Param        id  path  string  true  "Epic ID"
// @Success      204  "No Content"
// @Failure      404  {object}  api.ErrorResponse
// @Security     CookieAuth
// @Router       /api/epics/{id}/restore [post]
func (h *Handler) Restore(ctx context.Context) error {
	id := api.URLParam(ctx, "id")

	b, err := h.boardStore.GetByUser(ctx, auth.UserIDFromContext(ctx))
	if err != nil {
		return api.NotFound("board not found")
	}

	if err := h.store.RestoreEpic(ctx, id, b.ID); err != nil {
		return api.Internal("failed to restore epic")
	}
	return nil
}